Apply CR comments
